+++ src/ds/val.rs
@@ -272,7 +272,7 @@ impl<'a> Value<'a> {
     /// `Value` is a boolean value. Can be altered.
     ///
     /// # Example
-    ///	```rust
+    /// ```rust
     /// # use kserd::*;
     /// let mut value = Value::Bool(false);
     /// value.bool_mut().map(|x| *x = true);
@@ -340,7 +340,7 @@ impl<'a> Value<'a> {
     /// # Example
     /// ```rust
     /// # use kserd::*;
-    ///	let value = Value::new_num(-3.14);
+    /// let value = Value::new_num(-3.14);
     /// assert_eq!(value.float(), Some(-3.14));
     /// ```
     pub fn float(&self) -> Option<f64> {
@@ -388,7 +388,7 @@ impl<'a> Value<'a> {
     /// # Example
     /// ```rust
     /// # use kserd::*;
-    ///	let mut value = Value::new_str("Hello");
+    /// let mut value = Value::new_str("Hello");
     /// value.str_mut().map(|x| { x.push_str(", world!"); });
     /// assert_eq!(value.str(), Some("Hello, world!"));
     /// ```
@@ -436,38 +436,38 @@ impl<'a> Value<'a> {
 /// Conversions.
 impl<'a> Value<'a> {
     /// Clones all data to make a static `Value`.
-    pub fn to_owned(self) -> Value<'static> {
+    pub fn into_owned(self) -> Value<'static> {
         match self {
             Value::Unit => Value::Unit,
             Value::Bool(v) => Value::Bool(v),
             Value::Num(v) => Value::Num(v),
-            Value::Str(s) => Value::Str(s.to_owned()),
-            Value::Barr(b) => Value::Barr(b.to_owned()),
+            Value::Str(s) => Value::Str(s.into_owned()),
+            Value::Barr(b) => Value::Barr(b.into_owned()),
             Value::Tuple(seq) => Value::Tuple({
                 let mut v = Vec::with_capacity(seq.len());
                 for i in seq {
-                    v.push(i.to_owned())
+                    v.push(i.into_owned())
                 }
                 v
             }),
             Value::Cntr(map) => Value::Cntr({
                 let mut m = BTreeMap::new();
                 for (k, v) in map {
-                    m.insert(k.to_owned(), v.to_owned());
+                    m.insert(k.into_owned(), v.into_owned());
                 }
                 m
             }),
             Value::Seq(seq) => Value::Seq({
                 let mut v = Vec::with_capacity(seq.len());
                 for i in seq {
-                    v.push(i.to_owned())
+                    v.push(i.into_owned())
                 }
                 v
             }),
             Value::Map(map) => Value::Map({
                 let mut m = BTreeMap::new();
                 for (k, v) in map {
-                    m.insert(k.to_owned(), v.to_owned());
+                    m.insert(k.into_owned(), v.into_owned());
                 }
                 m
             }),
@@ -490,7 +490,7 @@ impl<'a> Value<'a> {
     /// ```
     ///
     /// [`Decoder`]: crate::encode::Decoder
-    pub fn mk_brw<'b>(&'b self) -> Value<'b> {
+    pub fn mk_brw(&self) -> Value {
         match &self {
             Value::Unit => Value::Unit,
             Value::Bool(v) => Value::Bool(*v),
@@ -534,7 +534,7 @@ impl<'a> Clone for Value<'a> {
         match &self {
             Value::Unit => Value::Unit,
             Value::Bool(v) => Value::Bool(*v),
-            Value::Num(v) => Value::Num(v.clone()),
+            Value::Num(v) => Value::Num(*v),
             Value::Str(s) => Value::Str(s.clone()),
             Value::Barr(b) => Value::Barr(b.clone()),
             Value::Tuple(seq) => Value::Tuple(seq.clone()),
@@ -685,5 +685,47 @@ mod tests {
             Kserd::new_num(3.14),
         ])));
         test2(Kserd::new_map(vec![(Kserd::new_unit(), Kserd::new_num(0))]));
+    }
+
+    #[test]
+    fn into_owned_test() {
+        use crate::to_kserd::ToKserd;
+
+        let prims = Kserd::new_cntr(vec![
+            ("unit", Kserd::new_unit()),
+            ("bool", Kserd::new_bool(true)),
+            ("num", Kserd::new_num(1.01234)),
+            ("str", Kserd::new_str("Hello, world!")),
+            ("barr", Kserd::new_barr([0, 1, 2, 4, 8].as_ref())),
+        ])
+        .unwrap();
+
+        let nested = Kserd::new_cntr(vec![
+            ("prims", prims.clone()),
+            (
+                "tuple",
+                Kserd::new(Value::Tuple(vec![prims.clone(), prims.clone()])),
+            ),
+            (
+                "seq",
+                Kserd::new(Value::Seq(vec![
+                    prims.clone(),
+                    prims.clone(),
+                    prims.clone(),
+                    prims.clone(),
+                ])),
+            ),
+            (
+                "map",
+                Kserd::new_map(vec![
+                    ("first".into_kserd().unwrap(), prims.clone()),
+                    ("second".into_kserd().unwrap(), prims.clone()),
+                ]),
+            ),
+        ])
+        .unwrap();
+
+        let nested_owned = nested.clone().into_owned();
+        assert_eq!(nested, nested_owned);
     }
 }