+++ src/parse/heuristics.rs
@@ -34,7 +34,7 @@ pub fn pattern_match_delimited<'a, E: ParseError<&'a str>>(
         }
     };
 
-    let (i, _) = multiline_wsp(i)?;
+    let (i, _) = multiline_whitespace(i)?;
     let next = i.chars().next().unwrap_or(' ');
 
     let mut r = if next == '[' {
@@ -45,7 +45,7 @@ pub fn pattern_match_delimited<'a, E: ParseError<&'a str>>(
         // opening of tuple or container
         if (ident_ == Ident::None || ident_ == Ident::Prim) && unit_value::<E>(i).is_ok() {
             Nonprim::None // a unit value, not a tuple. Can't be named, or would be an empty tuple
-        } else if recognise_field_assign(multiline_wsp(&i[1..])?.0) {
+        } else if recognise_field_assign(multiline_whitespace(&i[1..])?.0) {
             Nonprim::Cntr
         } else {
             Nonprim::Tuple
@@ -66,7 +66,7 @@ pub fn pattern_match_delimited<'a, E: ParseError<&'a str>>(
 pub fn pattern_match_verbose<'a, E: ParseError<&'a str>>(
     orig: &'a str,
 ) -> IResult<&'a str, bool, E> {
-    let (i, _) = multiline_wsp(orig)?; // ignore _all_ whitespace before hand
+    let (i, _) = multiline_whitespace(orig)?; // ignore _all_ whitespace before hand
 
     // generally a Verbose format can be recognised by:
     // - A direct mapping (var_name = xxx)
@@ -77,13 +77,19 @@ pub fn pattern_match_verbose<'a, E: ParseError<&'a str>>(
     let is = is
         || preceded::<_, _, _, (), _, _>(
             char('['),
-            terminated(ignore_inline_wsp(valid_name), ignore_inline_wsp(char(']'))),
+            terminated(
+                ignore_inline_whitespace(valid_name),
+                ignore_inline_whitespace(char(']')),
+            ),
         )(i)
         .is_ok();
     let is = is
         || preceded::<_, _, _, (), _, _>(
             tag("[["),
-            terminated(ignore_inline_wsp(valid_name), ignore_inline_wsp(tag("]]"))),
+            terminated(
+                ignore_inline_whitespace(valid_name),
+                ignore_inline_whitespace(tag("]]")),
+            ),
         )(i)
         .is_ok();
 
@@ -104,7 +110,7 @@ pub fn recognise_concise(i: &str) -> bool {
 
 /// Recognises the sequence: `field_name =`.
 pub fn recognise_field_assign(i: &str) -> bool {
-    preceded::<_, _, _, (), _, _>(field_name, terminated(inline_wsp, char('=')))(i).is_ok()
+    preceded::<_, _, _, (), _, _>(field_name, terminated(inline_whitespace, char('=')))(i).is_ok()
 }
 
 #[cfg(test)]
@@ -112,9 +118,7 @@ mod tests {
     use super::*;
     use nom::error::VerboseError;
 
-    #[test]
-    fn test_pattern_matching() {
-        macro_rules! test {
+    macro_rules! test {
 		( $ex:expr, $($x:literal ),* ) => {{
 			$(
 				let r = pattern_match_delimited::<VerboseError<_>>($x);
@@ -123,6 +127,8 @@ mod tests {
 		}};
 	}
 
+    #[test]
+    fn test_pat_match_none() {
         test!(
             Nonprim::None,
             "123456",
@@ -135,7 +141,10 @@ mod tests {
             "false",
             "b91':C!WEH#L4R'"
         );
+    }
 
+    #[test]
+    fn test_pat_match_seq() {
         test!(
             Nonprim::Seq,
             "[1,2,3]",
@@ -145,7 +154,10 @@ mod tests {
             "Hello[1,2,3]",
             "Hello[\n1,2,3\n]"
         );
+    }
 
+    #[test]
+    fn test_pat_match_map() {
         test!(
             Nonprim::Map,
             "{1:2, 3:4, 5:6}",
@@ -155,7 +167,10 @@ mod tests {
             "Hello{1:2, 3:4, 5:6}",
             "Hello{\n1:2\n3:4\n5:6\n}"
         );
+    }
 
+    #[test]
+    fn test_pat_match_tuple() {
         test!(
             Nonprim::Tuple,
             "(1,2,3)",
@@ -165,7 +180,10 @@ mod tests {
             "Hello(1,2,3)",
             "Hello(\n1,2,3\n)"
         );
+    }
 
+    #[test]
+    fn test_pat_match_cntr() {
         test!(
             Nonprim::Cntr,
             "(a=1, 2, 3)",