+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+# Version
+- Rename `ParseErr` to `Error`
+- Rename internal items to make clippy happy
+- Rename `to_owned` to `into_owned`
+
 # 0.2.0
 - Remove `From<String>` impl for `ToKserdErr`
 - Cleaned up error handling